Fixed the player character display bug found in the gameplay_loop where player 1's character was being printed out for each player.  Also got rid of the integer conversion for nodes_in_sequence_to_rotate inside the rotate method since the input obtainer for sub_board_length does this anyway.
Deleted the main file and moved the starting code into the ScalablePentago class.  The code will run if the class's file is run, and it won't effect others who use the class for their own means.

In addition, I deleted my message describing the reason why the abstract data type and it's implementation were not split for the class.  To put simply, it's not necessary to mention and it might bring confusion to those reading the code (plus, I need to learn more about interfaces before talking about them).